+++ frontend/classic/js/cm-provider-container.ts
@@ -107,9 +107,6 @@ export class ProviderContainer {
 
             scroll.destroy();
         })();
-
-        // Send the document creation request.
-        manager.coq.newDoc({uri: doc.getUri(), version: doc.getVersion(), raw: doc.getRawValue()})
     }
 
     /**